move android detection to core util
