+++ kolibri/plugins/learn/assets/src/views/content-page/index.vue
@@ -116,6 +116,7 @@
   import contentRenderer from 'kolibri.coreVue.components.contentRenderer';
   import downloadButton from 'kolibri.coreVue.components.downloadButton';
   import kButton from 'kolibri.coreVue.components.kButton';
+  import { isAndroidWebView } from 'kolibri.utils.browser';
   import assessmentWrapper from '../assessment-wrapper';
   import pointsPopup from '../points-popup';
   import pointsSlidein from '../points-slidein';
@@ -136,45 +137,12 @@
     },
     data: () => ({ wasIncomplete: false }),
     computed: {
-      /**
-        * Detects whether an Android device is using WebView.
-        * Based on https://developer.chrome.com/multidevice/user-agent#webview_user_agent
-        */
-      isAndroidWebView() {
-        const ua = window.navigator.userAgent;
-        const isAndroid = /Android/.test(ua);
-
-        if (isAndroid) {
-          const androidVersion = parseFloat(ua.match(/Android\s([0-9\.]*)/)[1]);
-          const isChrome = /Chrome/.test(ua);
-
-          // WebView UA in Lollipop and Above
-          // Android >=5.0
-          if (androidVersion >= 5.0 && isChrome && /wv/.test(ua)) {
-            return true;
-          }
-
-          // WebView UA in KitKat to Lollipop
-          // Android >= 4.4
-          if (androidVersion >= 4.4 && androidVersion < 5.0 && isChrome && /Version\//.test(ua)) {
-            return true;
-          }
-
-          // Old WebView UA
-          // Android < 4.4
-          if (androidVersion < 4.4 && /Version\//.test(ua) && /\/534.30/.test(ua)) {
-            return true;
-          }
-        }
-
-        return false;
-      },
       canDownload() {
         if (this.content) {
           return (
             this.downloadableFiles.length &&
             this.content.kind !== ContentNodeKinds.EXERCISE &&
-            !this.isAndroidWebView
+            !isAndroidWebView()
           );
         }
         return false;
